src/app/components/style-guide/typography/typography.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component, DebugElement } from '@angular/core';
import { By } from '@angular/platform-browser';
import { TypographyComponent } from './typography.component';

describe('Component: StyleGuidetypography', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [TypographyComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder): void {
    builder = tcb;
  }));

  it('should inject the component', inject([TypographyComponent], (component: TypographyComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(TypographyTestControllerComponent)
      .then((fixture: ComponentFixture<any>) => {
        let query: DebugElement = fixture.debugElement.query(By.directive(TypographyComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  directives: [TypographyComponent],
  selector: 'td-test',
  template: `
    <td-style-guide-typography></td-style-guide-typography>
  `,
})
class TypographyTestControllerComponent {
}
